pyproject.toml
[project]
name = "david-workshop"
version = "0.1.0"
description = "Discord bot for fandom events"
authors = ["NessaC <darklinaserverbot@gmail.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "sqlalchemy",
    "psycopg2-binary",
    "pytest",
    "pytest-asyncio",
    "pre-commit",
    "pillow"
]
[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:discord.player",
    "ignore::DeprecationWarning:audioop"
]
